pyproject.toml
[tool.poetry]
name = "repl_python3_Official-Discord-Leisure-Centre-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
psutil = "^5.8.0"
dta = "^1.2.1"
aiofiles = "^0.8.0"
"discord.py" = {git = "https://github.com/Rapptz/discord.py", extras = ["speed"]}
jishaku = "^2.4.0"
python-dotenv = "^0.20.0"
orjson = "^3.6.7"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
